♻️ refactor: ディレクトリ構成の変更 #9
