+++ pages/articles/[slug].jsx
@@ -1,23 +1,23 @@
 // 記事ページ
 
 import { getPostBySlug, getAllSlugs, getAllCategories } from '@/libs/api'
-import Container from '@/src/components/container/container'
-import PostHeader from '@/src/components/layouts/post-header/post-header'
+import Container from '@/src/components/layouts/container/Container'
+import PostHeader from '@/src/components/layouts/post-header/PostHeader'
 import Image from 'next/image'
 import {
   TwoColum,
   TwoColumMain,
   TwoColumSidebar,
 } from '@/src/components/layouts/two-colum/TwoColum'
-import ConvertBody from '@/src/components/convert/convert-body'
+import ConvertBody from '@/src/components/convert/ConvertBody'
 import PostCategories from '@/src/components/post-categories/post-categories'
 import { extractText } from '@/libs/extract-text'
-import Meta from '@/src/components/meta/meta'
+import Meta from '@/src/components/meta/Meta'
 import { prevNextPost } from '@/libs/prev-next-post'
-import Pagination from '@/src/components/pagination/pagination'
-import PostBody from '@/src/components/post-body/post-body'
+import Pagination from '@/src/components/pagination/Pagination'
+import PostBody from '@/src/components/post-body/PostBody'
 import { renderToc } from '@/libs/render-toc'
-import TableOfContents from '@/src/components/table-of-contents/table-of-contents'
+import TableOfContents from '@/src/components/table-of-contents/TableOfContents'
 import 'highlight.js/styles/night-owl.css'
 
 export default function Post({
