+++ pages/contact.jsx
@@ -1,8 +1,8 @@
 // contactページ
 
-import Meta from '@/src/components/meta/meta'
-import Container from '@/src/components/container/container'
-import Hero from '@/src/components/hero/hero'
+import Meta from '@/src/components/meta/Meta'
+import Container from '@/src/components/layouts/container/Container'
+import Hero from '@/src/components/hero/Hero'
 import Image from 'next/image'
 import eyecatch from 'images/contact.jpg'
 import Form from '@/src/components/form/form'
