+++ src/components/hero/hero.jsx
@@ -1,4 +1,4 @@
-import styles from './hero.module.css'
+import styles from './Hero.module.css'
 
 export default function Hero({ title, subtitle }) {
   return (
