+++ src/components/layouts/footer/Footer.jsx
@@ -1,5 +1,5 @@
 import styles from './Footer.module.css'
-import Social from '../../social/social'
+import Social from '../../social/Social'
 import Logo from '../../logo/logo'
 
 export default function Footer() {
