+++ src/components/layouts/post-header/PostHeader.jsx
@@ -1,9 +1,9 @@
 // 記事のヘッダー部分 (タイトル、サブタイトル、投稿日)
 
-import styles from './post-header.module.css'
+import styles from './PostHeader.module.css'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faClock } from '@fortawesome/free-regular-svg-icons'
-import ConvertDate from '../../convert/convert-date'
+import ConvertDate from '../../convert/ConvertDate'
 import Image from 'next/image'
 
 export default function PostHeader({ icon, title, subtitle, publish = '' }) {
