+++ src/components/nav/nav.jsx
@@ -1,7 +1,7 @@
 // ナビコンポーネント
 
 import Link from 'next/link'
-import styles from './nav.module.css'
+import styles from './Nav.module.css'
 import { useState } from 'react'
 
 export default function Nav() {
