+++ src/components/post-body/PostBody.jsx
@@ -1,6 +1,6 @@
 // 記事ページのbodyコンポーネント
 
-import styles from './post-body.module.css'
+import styles from './PostBody.module.css'
 
 export default function PostBody({ children }) {
   return <div className={styles.stack}>{children}</div>
